moved to utilities. edited changelog
